Documentation added for Word Splitting and Typo Replacement
