+++ docs/code/data_aug.rst
@@ -73,6 +73,16 @@ Data Augmentation Ops
 :hidden:`WordnetDictionary`
 ----------------------------
 .. autoclass:: forte.processors.data_augment.algorithms.dictionary.WordnetDictionary
+    :members: 
+
+:hidden:`TypoReplacementOp`
+----------------------------
+.. autoclass:: forte.processors.data_augment.algorithms.typo_replacement_op.TypoReplacementOp
+    :members:
+
+:hidden:`WordSplittingOp`
+----------------------------
+.. autoclass:: forte.processors.data_augment.algorithms.word_splitting_processor.RandomWordSplitDataAugmentProcessor
     :members:
 
 Data Augmentation Models
