+++ forte/processors/misc/simple_processors.py
@@ -15,12 +15,11 @@ import re
 
 from forte.data import DataPack
 from forte.processors.base import PackProcessor
-from ft.onto.base_ontology import EntityMention, Sentence, Token
+from ft.onto.base_ontology import Sentence, Token
 
 __all__ = [
     "PeriodSentenceSplitter",
-    "WhiteSpaceTokenizer",
-    "EntityMentionInserter",
+    "WhiteSpaceTokenizer"
 ]
 
 
@@ -59,30 +58,3 @@ class WhiteSpaceTokenizer(PackProcessor):
 
         if start < len(input_pack.text):
             input_pack.add_entry(Token(input_pack, start, len(input_pack.text)))
-
-
-class EntityMentionInserter(PackProcessor):
-    """
-    A simple processor that inserts Entity Mentions into the data pack.
-    The input required is the annotations that wish to be tagged as Entity
-    Mentions. If the given annotations are not present in the given data pack,
-    an exception is raised.
-    """
-
-    def _process(self, input_pack: DataPack):
-        entity_text = self.configs.entities_to_insert
-
-        input_text = input_pack.text
-        if not all(bool(entity in input_text) for entity in entity_text):
-            raise Exception(
-                "Entities to be added are not valid for the input text."
-            )
-        for entity in entity_text:
-            start = input_text.index(entity)
-            end = start + len(entity)
-            entity_mention = EntityMention(input_pack, start, end)
-            input_pack.add_entry(entity_mention)
-
-    @classmethod
-    def default_configs(cls):
-        return {"entities_to_insert": []}
